+++ src/components/utility/NodeTree.test.tsx
@@ -1,5 +1,6 @@
-﻿import ResumeNodeTree, { BasicResumeNode } from "./NodeTree";
+﻿import ResumeNodeTree from "./NodeTree";
 import { assignIds } from "../Helpers";
+import { BasicResumeNode } from "./Types";
 
 test('getNodeById Test', () => {
     const resumeData = [{
